requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=5.0.0

# Google Trends
pytrends>=4.9.0

# AI - Gemini
google-genai>=1.0.0

# Google Sheets
gspread>=6.0.0
google-auth>=2.20.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
